requirements.txt
django>=5.0.0
djangorestframework>=3.14.0
transformers>=4.36.0
torch>=2.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

